Add ecommerce app for SA/SE bootcamp
