+++ ecommerce_server_python/requirements.txt
@@ -1,0 +1,4 @@
+fastmcp>=0.1.0
+mcp>=0.1.0
+pydantic>=2.6.0
+uvicorn>=0.30.0
